Factor out auth into middleware
