+++ server/src/app.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import pino from 'pino';
 import pinoHttp from 'pino-http';
 import cors from 'cors';
-import root from './root';
+import {root} from './root';
 
 const PORT = 8000;
 const logger = pino();
@@ -21,7 +21,7 @@ app.use(httpLogger);
 app.use(express.json());
 app.use(cors());
 
-app.use("/", root);
+app.use('/', root);
 
 app.listen(PORT, () => {
   logger.info(`Listening on port ${PORT}...`);
