+++ server/tsconfig.json
@@ -5,7 +5,8 @@
     "rootDir": "src",
     "outDir": "build",
     "esModuleInterop": true,
-    "strict": true
+    "strict": true,
+    "typeRoots": ["./types", "./node_modules/@types"]
   },
   "include": ["src"],
   "exclude": ["build"]
