+++ server/types/express/index.d.ts
@@ -1,0 +1,5 @@
+declare namespace Express {
+    export interface Request {
+        sub: string
+    }
+}
